add delete and status feature to items

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -98,11 +99,16 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
+
+
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $variant->id)->count();
+
+        if($variantItemCheck > 0) {
+            return response(['status' => 'error', 'message' => 'This variant  contains items in it! Delete this item first!']);
+        }
+
         $variant->delete();
-
-
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
-
     }
 
     public function changeStatus(Request $request){
